Use a more consistent search pattern
The whole point of our module is to offer to use another value than
JHipster provided one. So when regenerating, depend on what the user do
(letting JHipster overwrite or change or not) we might want to search
for JHipster value or our module value. So we search for both with a
regex alternative.